+++ cloudbridge/__init__.py
@@ -2,12 +2,12 @@ import logging
 import sys
 
 # Current version of the library
-__version__ = '0.1.0'
+__version__ = '0.1.1'
 
 
 def get_version():
     """
-    Returns a string with the current version of the library (e.g., "0.1.0")
+    Return a string with the current version of the library (e.g., "0.1.0").
     """
     return __version__