+++ cloudbridge/cloud/providers/gce/resources.py
@@ -17,6 +17,7 @@ except NameError:
 import hashlib
 import inspect
 import json
+import re
 
 class GCEKeyPair(BaseKeyPair):
 
@@ -176,49 +177,80 @@ class GCERegion(BaseRegion):
 
 
 class GCEFirewallsDelegate(object):
+    DEFAULT_NETWORK = 'default'
+    _NETWORK_URL_PREFIX = 'global/networks/'
   
     def __init__(self, provider):
         self._provider = provider
         self._list_response = None
 
     @staticmethod
-    def tag_id(tag):
+    def tag_network_id(tag, network):
+        """
+        Generate an ID for a (tag, network) pair.
+        """
         md5 = hashlib.md5()
-        md5.update(tag.encode('ascii'))
+        md5.update("{0}-{1}".format(tag, network).encode('ascii'))
         return md5.hexdigest()
+
+    @staticmethod
+    def network(firewall):
+        """
+        Extract the network name of a firewall.
+        """
+        if 'network' not in firewall:
+            return GCEFirewallsDelegate.DEFAULT_NETWORK
+        match = re.search(
+                GCEFirewallsDelegate._NETWORK_URL_PREFIX + '([^/]*)$',
+                firewall['network'])
+        if match and len(match.groups()) == 1:
+            return match.group(1)
+        return None
 
     @property
     def provider(self):
         return self._provider
 
     @property
-    def tags(self):
+    def tag_networks(self):
+        """
+        List all (tag, network) pairs that are used in at least one firewall.
+        """
         out = set()
         for firewall in self.iter_firewalls():
-            out.add(firewall['targetTags'][0])
+            network = GCEFirewallsDelegate.network(firewall)
+            if network is not None:
+                out.add((firewall['targetTags'][0], network))
         return out
             
-    def get_tag_from_id(self, tag_id):
-        for tag in self.tags:
-            if GCEFirewallsDelegate.tag_id(tag) == tag_id:
-                return tag
-        return None
+    def get_tag_network_from_id(self, tag_network_id):
+        """
+        Map an ID back to the (tag, network) pair.
+        """
+        for tag, network in self.tag_networks:
+            current_id = GCEFirewallsDelegate.tag_network_id(tag, network)
+            if current_id == tag_network_id:
+                return (tag, network)
+        return (None, None)
 
-    def has_tag(self, tag):
-        return tag in self.tags
-
-    def delete_tag_with_id(self, tag_id):
-        tag = self.get_tag_from_id(tag_id)
+    def delete_tag_network_with_id(self, tag_network_id):
+        """
+        Delete all firewalls in a given network with a specific target tag.
+        """
+        tag, network = self.get_tag_network_from_id(tag_network_id)
         if tag is None:
             return
-        for firewall in self.iter_firewalls(tag):
+        for firewall in self.iter_firewalls(tag, network):
             self._delete_firewall(firewall)
         self._update_list_response()
 
     def add_firewall(self, tag, ip_protocol, port, source_range, source_tag,
-                     description):
+                     description, network):
+        """
+        Create a new firewall.
+        """
         if self.find_firewall(tag, ip_protocol, port, source_range,
-                              source_tag) is not None:
+                              source_tag, network) is not None:
             return True
         # Do not let the user accidentally open traffic from the world by not
         # explicitly specifying the source.
@@ -226,16 +258,18 @@ class GCEFirewallsDelegate(object):
             return False
         firewall_number = 1
         suffixes = []
-        for firewall in self.iter_firewalls(tag):
+        for firewall in self.iter_firewalls(tag, network):
             suffix = firewall['name'].split('-')[-1]
             if suffix.isdigit():
                 suffixes.append(int(suffix))
         for suffix in sorted(suffixes):
             if firewall_number == suffix:
                 firewall_number += 1
-        firewall = {'name': '%s-rule-%d' % (tag, firewall_number),
-                    'allowed': [{'IPProtocol': str(ip_protocol)}],
-                    'targetTags': [tag]}
+        firewall = {
+            'name': '%s-%s-rule-%d' % (network, tag, firewall_number),
+            'network': GCEFirewallsDelegate._NETWORK_URL_PREFIX + network,
+            'allowed': [{'IPProtocol': str(ip_protocol)}],
+            'targetTags': [tag]}
         if description is not None:
             firewall['description'] = description
         if port is not None:
@@ -259,10 +293,14 @@ class GCEFirewallsDelegate(object):
         finally:
             self._update_list_response()
 
-    def find_firewall(self, tag, ip_protocol, port, source_range, source_tag):
+    def find_firewall(self, tag, ip_protocol, port, source_range, source_tag,
+                      network):
+        """
+        Find a firewall with give parameters.
+        """
         if source_range is None and source_tag is None:
             source_range = '0.0.0.0/0'
-        for firewall in self.iter_firewalls(tag):
+        for firewall in self.iter_firewalls(tag, network):
             if firewall['allowed'][0]['IPProtocol'] != ip_protocol:
                 continue
             if not self._check_list_in_dict(firewall['allowed'][0], 'ports',
@@ -277,6 +315,9 @@ class GCEFirewallsDelegate(object):
         return None
 
     def get_firewall_info(self, firewall_id):
+        """
+        Extract firewall properties to into a dictionary for easy of use.
+        """
         info = {}
         for firewall in self.iter_firewalls():
             if firewall['id'] != firewall_id:
@@ -293,16 +334,24 @@ class GCEFirewallsDelegate(object):
             if ('ports' in firewall['allowed'][0] and
                 len(firewall['allowed'][0]['ports']) == 1):
                 info['port'] = firewall['allowed'][0]['ports'][0]
+            info['network'] = GCEFirewallsDelegate.network(firewall)
             return info
         return info
 
     def delete_firewall_id(self, firewall_id):
+        """
+        Delete a firewall with a given ID.
+        """
         for firewall in self.iter_firewalls():
             if firewall['id'] == firewall_id:
                 self._delete_firewall(firewall)
         self._update_list_response()
 
-    def iter_firewalls(self, tag=None):
+    def iter_firewalls(self, tag=None, network=None):
+        """
+        Iterate through all firewalls. Can optionally iterate through firewalls
+        with a given tag and/or in a network.
+        """
         if self._list_response is None:
             self._update_list_response()
         if 'items' not in self._list_response:
@@ -312,10 +361,19 @@ class GCEFirewallsDelegate(object):
                 continue
             if 'allowed' not in firewall or len(firewall['allowed']) != 1:
                 continue
-            if tag is None or firewall['targetTags'][0] == tag:
+            if tag is not None and firewall['targetTags'][0] != tag:
+                continue
+            if network is None:
+                yield firewall
+                continue
+            firewall_network = GCEFirewallsDelegate.network(firewall)
+            if firewall_network == network:
                 yield firewall
 
     def _delete_firewall(self, firewall):
+        """
+        Delete a given firewall.
+        """
         project_name = self._provider.project_name
         try:
             response = (self._provider.gce_compute
@@ -330,6 +388,9 @@ class GCEFirewallsDelegate(object):
             return False
 
     def _update_list_response(self):
+        """
+        Sync the local cache of all firewalls with the server.
+        """
         self._list_response = (
                 self._provider.gce_compute
                               .firewalls()
@@ -337,6 +398,9 @@ class GCEFirewallsDelegate(object):
                               .execute())
 
     def _check_list_in_dict(self, dictionary, field_name, value):
+        """
+        Verify that a given field in a dictionary is a singlton list [value].
+        """
         if field_name not in dictionary:
             return value is None
         if (value is None or
@@ -348,24 +412,46 @@ class GCEFirewallsDelegate(object):
 
 class GCESecurityGroup(BaseSecurityGroup):
 
-    def __init__(self, delegate, tag, description=None):
+    def __init__(self, delegate, tag,
+                 network=GCEFirewallsDelegate.DEFAULT_NETWORK,
+                 description=None):
         super(GCESecurityGroup, self).__init__(delegate.provider, tag)
         self._description = description
         self._delegate = delegate
+        self._network = network
+        if self._network is None:
+            self._network = GCEFirewallsDelegate.DEFAULT_NETWORK 
 
     @property
     def id(self):
-        return GCEFirewallsDelegate.tag_id(self._security_group)
+        """
+        Return the ID of this security group which is determined based on the
+        network and the target tag corresponding to this security group.
+        """
+        return GCEFirewallsDelegate.tag_network_id(self._security_group,
+                                                   self._network)
 
     @property
     def name(self):
+        """
+        Return the name of the security group which is the same as the
+        corresponding tag name.
+        """
         return self._security_group
 
     @property
     def description(self):
+        """
+        The description of the security group is even explicitly given when the
+        group is created or is determined from a firewall in the group.
+
+        If the firewalls are created using this API, they all have the same
+        description.
+        """
         if self._description is not None:
             return self._description
-        for firewall in self._delegate.iter_firewalls(self._security_group):
+        for firewall in self._delegate.iter_firewalls(self._security_group,
+                                                      self._network):
             if 'description' in firewall:
                 return firewall['description']
         return None
@@ -373,7 +459,8 @@ class GCESecurityGroup(BaseSecurityGroup):
     @property
     def rules(self):
         out = []
-        for firewall in self._delegate.iter_firewalls(self._security_group):
+        for firewall in self._delegate.iter_firewalls(self._security_group,
+                                                      self._network):
             out.append(GCESecurityGroupRule(self._delegate, firewall['id']))
         return out
 
@@ -391,7 +478,8 @@ class GCESecurityGroup(BaseSecurityGroup):
         port = GCESecurityGroup.to_port_range(from_port, to_port)
         src_tag = src_group.name if src_group is not None else None
         self._delegate.add_firewall(self._security_group, ip_protocol, port,
-                                    cidr_ip, src_tag, self.description)
+                                    cidr_ip, src_tag, self.description,
+                                    self._network)
         return self.get_rule(ip_protocol, from_port, to_port, cidr_ip,
                              src_group)
 
@@ -400,7 +488,8 @@ class GCESecurityGroup(BaseSecurityGroup):
         port = GCESecurityGroup.to_port_range(from_port, to_port)
         src_tag = src_group.name if src_group is not None else None
         firewall_id = self._delegate.find_firewall(
-                self._security_group, ip_protocol, port, cidr_ip, src_tag)
+                self._security_group, ip_protocol, port, cidr_ip, src_tag,
+                self._network)
         if firewall_id is None:
             return None
         return GCESecurityGroupRule(self._delegate, firewall_id)
@@ -426,10 +515,14 @@ class GCESecurityGroupRule(BaseSecurityGroupRule):
 
     @property
     def parent(self):
+        """
+        Return the security group to which this rule belongs.
+        """
         info = self._delegate.get_firewall_info(self._rule)
-        if info is None or 'target_tag' not in info:
+        if info is None or 'target_tag' not in info or info['network'] is None:
             return None
-        return GCESecurityGroup(self._delegate, info['target_tag'])
+        return GCESecurityGroup(self._delegate, info['target_tag'],
+                                info['network'])
 
     @property
     def id(self):
@@ -474,6 +567,9 @@ class GCESecurityGroupRule(BaseSecurityGroupRule):
 
     @property
     def cidr_ip(self):
+        """
+        Return the IP of machines from which this rule allows traffic.
+        """
         info = self._delegate.get_firewall_info(self._rule)
         if info is None or 'source_range' not in info:
             return None
@@ -481,10 +577,14 @@ class GCESecurityGroupRule(BaseSecurityGroupRule):
 
     @property
     def group(self):
+        """
+        Return the security group from which this rule allows traffic.
+        """
         info = self._delegate.get_firewall_info(self._rule)
-        if info is None or 'source_tag' not in info:
+        if info is None or 'source_tag' not in info or info['network'] is None:
             return None
-        return GCESecurityGroup(self._delegate, info['source_tag'])
+        return GCESecurityGroup(self._delegate, info['source_tag'],
+                                info['network'])
 
     def to_json(self):
         attr = inspect.getmembers(self, lambda a: not(inspect.isroutine(a)))