+++ CHANGELOG.rst
@@ -1,3 +1,14 @@
+0.1.1 - Aug 10, 2016.
+-------
+
+* For AWS, always launch instances into private networking (i.e., VPC)
+* Support for using OpenStack Keystone v3
+* Add functionality to manipulate routers and routes
+* Add FloatingIP resource type and integrate with Network service
+* Numerous documentation updates
+* For an OpenStack provider, add method to get the ec2 credentials for a user
+
+
 0.1.0 - Jan 30, 2016.
 -------